Sprites now follow the same rotation principles as masks so they can be "synched" up. Next up, Programming the new mask collisions api. Also, programming a shortcut method that checks mask's width and height and only performs complex collisions if the bounding boxes are colliding.

+++ src/gameengine/View.java
@@ -48,6 +48,8 @@ public class View {
         
         clearDisplay(graphics);
         
+        
+        
         GameObject.setOffset(new Point(roomX, roomY));
         GameObject.setZoom(xScale, yScale);
         drawGameObjects(graphics);
@@ -103,12 +105,12 @@ public class View {
         this.h = h;
     }
     
-    public void setXScale(int x)
+    public void setXScale(float x)
     {
         this.xScale = x;
     }
     
-    public void setYScale(int y)
+    public void setYScale(float y)
     {
         this.yScale = y;
     }
